refactor(12factor-env): move dotenv to a node-only entry so the main entry stays bundler-safe
