+++ packages/12factor-env/src/index.ts
@@ -1,7 +1,3 @@
-import { readFileSync } from 'node:fs';
-import path from 'node:path';
-import { parseEnv as parseEnvSource } from 'node:util';
-
 import type { CleanedEnv, CleanOptions, Spec,ValidatorSpec } from 'envalid';
 import {
   bool,
@@ -194,61 +190,6 @@ const boolish = makeValidator<boolean>((value: string) => {
   if (typeof raw === 'boolean') return raw;
   return parseEnvBoolean(String(raw)) ?? false;
 });
-
-// ── Opt-in dotenv support ────────────────────────────────────────────────────
-//
-// The 12-factor rule is "environment first, `.env` as a local-dev convenience":
-// a container gets its configuration from the process environment, while a
-// developer's project folder may carry a `.env`. `dotenv()` produces an
-// environment record that honors that rule — file values fill gaps, real
-// environment variables always win (unless `override` is set). It is a pure
-// input builder for `env()`/`cleanEnv`; it never mutates `process.env`.
-
-export type DotenvOptions = {
-  /** Explicit path to the env file. Takes precedence over `cwd`/`file`. */
-  path?: string;
-  /** Directory to resolve `file` in. Default: `process.cwd()`. */
-  cwd?: string;
-  /** File name resolved under `cwd`. Default: `.env`. */
-  file?: string;
-  /** Base environment the file is merged into. Default: `process.env`. */
-  environment?: Record<string, string | undefined>;
-  /** When true, file values win over the base environment. Default: false. */
-  override?: boolean;
-};
-
-/** Parse dotenv-format source text into a plain record (no interpolation). */
-export const parseDotenv = (source: string): Record<string, string> =>
-  parseEnvSource(source) as Record<string, string>;
-
-/**
- * Read an env file and merge it with an environment record. A missing file is
- * not an error — the base environment is returned unchanged, so the same code
- * path works in a container (no file) and in local dev (file present).
- */
-export const dotenv = (
-  options: DotenvOptions = {}
-): Record<string, string | undefined> => {
-  const {
-    environment = process.env,
-    override = false,
-    cwd = process.cwd(),
-    file = '.env'
-  } = options;
-  const filePath = options.path ?? path.join(cwd, file);
-
-  let source: string;
-  try {
-    source = readFileSync(filePath, 'utf8');
-  } catch {
-    return { ...environment };
-  }
-
-  const fileVars = parseDotenv(source);
-  return override
-    ? { ...environment, ...fileVars }
-    : { ...fileVars, ...environment };
-};
 
 // Type for specs object
 type Specs = Record<string, ValidatorSpec<unknown>>;
